// src/pages/annotator/tools/polygon-tool/polygon-tool.enum.ts
// INTEL CONFIDENTIAL
//
// Copyright (C) 2021 Intel Corporation
//
// This software and the related documents are Intel copyrighted materials, and your use of them is governed by
// the express license under which they were provided to you ("License"). Unless the License provides otherwise,
// you may not use, modify, copy, publish, distribute, disclose or transmit this software or the related documents
// without Intel's prior written permission.
//
// This software and the related documents are provided as is, with no express or implied warranties,
// other than those that are expressly stated in the License.

export enum PolygonMode {
    Polygon = 'Polygon',
    Lasso = 'Lasso',
    Eraser = 'Eraser',
    MagneticLasso = 'MagneticLasso',
}

export enum PointerIcons {
    Polygon = 'polygon-tool',
    Lasso = 'lasso-tool',
    Eraser = 'eraser-tool',
    MagneticLasso = 'magnetic-lasso-tool',
}

export enum PointerIconsOffset {
    Polygon = '10 25',
    Lasso = '10 25',
    Eraser = '15 15',
    MagneticLasso = '10 25',
}